Switch to DB-First, introduce ManufacturerIdBM and CategoryIdBM, update API

+++ TomiSoft.ProductCatalog/BusinessModels/ProductBM.cs
@@ -3,7 +3,7 @@ using TomiSoft.ProductCatalog.BusinessModels.Concepts;
 
 namespace TomiSoft.ProductCatalog.BusinessModels {
     public class ProductBM {
-        public ProductBM(BarcodeBM barcode, int? manufacturerId, int? categoryId, IReadOnlyDictionary<string, string> productName) {
+        public ProductBM(BarcodeBM barcode, long? manufacturerId, long? categoryId, IReadOnlyDictionary<string, string> productName) {
             Barcode = barcode;
             ManufacturerId = manufacturerId;
             CategoryId = categoryId;
@@ -11,8 +11,8 @@ namespace TomiSoft.ProductCatalog.BusinessModels {
         }
 
         public BarcodeBM Barcode { get; }
-        public int? ManufacturerId { get; }
-        public int? CategoryId { get; }
+        public long? ManufacturerId { get; }
+        public long? CategoryId { get; }
         public IReadOnlyDictionary<string, string> ProductName { get; }
     }
 }
